// src/types/ListType.ts
export type ListType = {
    id: number;
    title: string;
    nameLink1?: string;
    nameLink2?: string;
    nameLink3?: string;
    nameLink4?: string;
    nameLink5?: string;
    nameLink6?: string;
    nameLink7?: string;
    nameLink8?: string;
    link1?: string;
    link2?: string;
    link3?: string;
    link4?: string;
    link5?: string;
    link6?: string;
    link7?: string;
    link8?: string;
}